Explore control node done

+++ Limo/ros-packages/packages/src/explore_control/src/control_explore.py
@@ -4,15 +4,16 @@ import rospy
 from std_msgs.msg import Bool
 import os
 import subprocess
+from explore_control.msg import BoolString
 
 class ExplorationControl:
     def __init__(self):
         self.isSimulation = os.environ.get("IS_SIMULATION")
-	self.bringup_path = os.environ.get("BRINGUP_PATH")
         if self.isSimulation : 
             self.limoId = os.environ.get("LIMO_ID", "1")
             rospy.init_node('exploration_control' + self.limoId)
             self.subscriberState = rospy.Subscriber(f"limo{self.limoId}/exploration_state", Bool, self.setExplorationState)
+            self.publishState = rospy.Publisher('/exploration_state_sim', BoolString, queue_size=10)
         else:
             rospy.init_node('exploration_control')
             self.subscriberState = rospy.Subscriber(
@@ -32,22 +33,29 @@ class ExplorationControl:
         if self.explore_lite_process is None:
             if self.isSimulation:
                 rospy.loginfo("Launching explore_control for simulated limo")
-                self.explore_lite_process = subprocess.Popen(
-                    ["roslaunch", f'{self.bringup_path}', "one_exploration.launch", f'ns:=/limo{self.limoId}', f'id:=limo{self.limoId}'],
-                    stderr=subprocess.PIPE, preexec_fn=os.setpgrp)
+                new_msg = BoolString()
+                new_msg.data = True
+                new_msg.info = f'{self.limoId}'
+                self.publishState.publish(new_msg)
             else:
                 rospy.loginfo("Launching explore_control for physical limo")
                 self.explore_lite_process = subprocess.Popen(
-                    ["roslaunch", f'{self.bringup_path}', "one_exploration.launch"],
+                    ["roslaunch", 'limo_bringup', "one_exploration.launch"],
                     stderr=subprocess.PIPE, preexec_fn=os.setpgrp)
-            self.warning_filter_process = subprocess.Popen(
-                ["grep", "-v", "TF_REPEATED_DATA", "buffer_core"],
-                stdin=self.explore_lite_process.stderr)
+                self.warning_filter_process = subprocess.Popen(
+                    ["grep", "-v", "TF_REPEATED_DATA", "buffer_core"],
+                    stdin=self.explore_lite_process.stderr)
 
     def stop_explore_lite(self):
-        if self.explore_lite_process is not None:
-            self.explore_lite_process.terminate()
-            self.explore_lite_process = None
+        if self.isSimulation:
+            new_msg = BoolString()
+            new_msg.data = False
+            new_msg.info = f'{self.limoId}'
+            self.publishState.publish(new_msg)
+        else:
+            if self.explore_lite_process is not None:
+                self.explore_lite_process.terminate()
+                self.explore_lite_process = None
 
 if __name__ == '__main__':
     ec = ExplorationControl()
